relocate all js files into assets and refactor algolia autocomplete into module
