+++ assets/css/components/_algolia-autocomplete.scss
