+++ assets/js/app.js
@@ -12,4 +12,16 @@ import 'bootstrap'; // adds functions to jQuery
 // uncomment if you have legacy code that needs global variables
 // global.$ = $;
 
+import './components/article_show';
+import './components/article_form';
+import './components/report_form';
+import autocomplete from './components/algolia-autocomplete';
+
 $('.dropdown-toggle').dropdown();
+
+$(document).ready(function() {
+    const $autocomplete = $('.js-user-autocomplete');
+    if (!$autocomplete.is(':disabled')) {
+        autocomplete($autocomplete, 'users', 'email')
+    }
+});
