+++ assets/js/components/article_show.js
@@ -1,8 +1,9 @@
+import $ from 'jquery';
 $(document).ready(function() {
     $('.js-like-article').on('click', function(e) {
         e.preventDefault();
         let $link = $(e.currentTarget);
-        $link.toggleClass('fa-heart-o').toggleClass('fa-heart');
+        $link.toggleClass('far').toggleClass('fas');
 
         $.ajax({
             method: 'POST',
