+++ package.json
@@ -2,6 +2,7 @@
     "devDependencies": {
         "@fortawesome/fontawesome-free": "^5.13.0",
         "@symfony/webpack-encore": "^0.30.0",
+        "autocomplete.js": "^0.37.1",
         "bootstrap": "^4.5.0",
         "core-js": "^3.0.0",
         "jquery": "^3.5.1",
