+++ templates/article/update.html.twig
@@ -2,11 +2,6 @@
 
 {% block title %}Edit article: {{ form.vars.value.title }}{% endblock %}
 
-{% block javascripts %}
-    {{ parent() }}
-    <script src="{{ asset('js/article_form.js') }}"></script>
-{% endblock %}
-
 {% block admin_body %}
     <h1>Edit this article</h1>
     {{ include('article/_form.html.twig', {
