+++ templates/base.html.twig
@@ -82,7 +82,7 @@
                     </a>
                 </li>
             </ul>
-            <span class="text-muted">Made with <span class="fa fa-heart" style="color: red;"></span> by Ereshkidal</span>
+            <span class="text-muted">Made with <span class="fas fa-heart" style="color: red;"></span> by Ereshkidal</span>
         </div>
     </footer>
         {% block javascripts %}
